pyproject.toml
[project]
name = "cardiverse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adjusttext>=1.3.0",
    "ipdb>=0.13.13",
    "llama-index>=0.13.6",
    "matplotlib>=3.10.3",
    "networkx>=3.4.2",
    "openai>=1.106.1",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "retrying>=1.4.2",
    "tqdm>=4.67.1",
    "tree-sitter==0.21.3",
    "tree-sitter-languages>=1.10.2",
    "tree-sitter-python==0.23.5",
]
